seperate gameChess
